refactor(background): extract 83 message handlers into domain-specific modules
Move message handling from the service-worker god-object into five dedicated handler modules (settings, search, ollama, diagnostics, commands), wired through a two-registry composition root (preInit/postInit). Service-worker.ts drops from ~2100 to ~470 lines.

Made-with: Cursor
